Significantly overhaul --print-type
--print-type now performs some basic common sub-expression lifting before
pretty printing the type to STDOUT. We now try and factor out some common types,
such as OperatingSystem, Arch, Compiler, along with the "meat" of a package -
the Library, Executable, Benchmark, and so on.

--print-type can also be given the name of any let-bound type produced by
`--print-type` (with no arguments) to "zoom in" on the definition of a
particular type. For example:

Main> :main --print-type RepoKind
    let RepoKind =
          < RepoHead : {} | RepoKindUnknown : { _1 : Text } | RepoThis : {} >

in  RepoKind

Related to #22.