Remove the use of a custom context (#28)
dhall-to-cabal previously used a custom context that contained the `v` function, the `VersionRange` type, and constructors for `VersionRange`. However, @Gabriel439 Gabriel has demonstrated that this is both unnecessary and ultimately leads to a *worse* user experience, as dhall-to-cabal can no longer interoperate with any other Dhall tool.

In this commit, VersionRange has been encoded directly in Dhall. This technique encodes all constructors as function arguments, and makes VersionRange a function of all possible constructors. The same trick is done with Version - a Version is exactly something that is made with `v : Text -> Version`.