Change guarding to be a top level function per component
This commit changes the guarding logic from:

  List { guard : Config -> Bool, body : A }

to simply

  Config -> A

This is achieved through the new ConfigTree type.

ConfigTree takes the unsaturated configuration lambda, and applies it with a
known variable name. Next, we find all uses of this variable w.r.t. config field
access and application. We then replace all occurances of these applications
with either True or False, building a tree of decisions as we go - this is the
ConfigTree.

With a ConfigTree extracted, it's trivial to then turn this into a
Cabal.CondTree.

This currently produces some verbose Cabal, but it's not going to be much work
to lift common elements up the tree and produce miminal Cabal files. This will
come in future work.