Alternative approach where ConfigTree is a monad
Now rather than finding all uses of an expression, we simply find an expression,
and replace it with a branch of all possible outcomes. The monad sequences the
possible expansions together to produce an exhaustive tree of possibilities.

The de-duplication is then performed as post processing when converting these
expression trees into Cabal configurations.