Change guarding to be a top level function per component (#31)
This commit changes the guarding logic from:

  List { guard : Config -> Bool, body : A }

to simply

  Config -> A

This is achieved through the new ConfigTree type.

ConfigTree takes the unsaturated configuration lambda, and applies it with a
known variable name. Next, we find all uses of this variable w.r.t. config field
access and application. We then replace all occurances of these applications
with either True or False, building a tree of decisions as we go - this is the
ConfigTree.

With a ConfigTree extracted, it's trivial to then turn this into a
Cabal.CondTree. This is post processed, with impossible conditions removed
and duplication pushed up as much as possible to avoid repetition.