Add a default version number to defaults/Package.dhall.
There's already a default empty 'name' field, and this isn't worse
than that. It means that the default Package has the type from
types/Package.dhall, which is a win for short examples and hopefully
will reduce user surprise--I was surprised, and the type error was
far too enormous to work out what was going on.

Also removed an incidental trailing bit of whitespace.