Fix tracking which branches are already true or false in conditionals (#56)
* Fix tracking which branches are already true or false in conditionals

Previously, only one list of already-seen conditions was kept, and
this was assumed to be the list of 'true' conditions. But that was
also being added to in the false branch! We actually need to keep two
lists, one of already-true and one of already-false.

Fixes #53. Fixes #55.

* Test that compiler options don't get reordered in conditionals.

* Test for conditional dependencies.

Since these are represented specially in Cabal's CondTree type, this
is worth testing explicitly.

* Update Changelog