cabal-to-dhall: keep the generated imports from GitHub up to date
This will grab the current version number from Paths_dhall_to_cabal
and then generate the URL to import from from that.

Further, now CabalToDhall.cabalToDhall is parameterised by the
location of the prelude.dhall and types.dhall to import.

As a knock-on change, to prevent the tests breaking every time the
version is bumped, these now include only relative paths.