Move things that aren't types out of dhall/types (#94)
* Move things that aren't types out of dhall/types

Mostly, these are the functions operating on the VersionRange and SPDX
types. The files have also been renamed to lowercase to better-reflect
their names in the prelude.

* Changelog for the reorganisation of dhall/

This is out of all proportion to the likely impact of the change
considering dhall-to-cabal's current userbase, but properly
documenting breaking changes is a good habit to keep.