Change the default build-type to omission and cabal-version to 2.2 (#98)
In Cabal 2.2, build-type inference works for 99% of cases and it is
pretty nice. However, if we have a specified default build-type,
there's no good way to get at that niceness from dhall-to-cabal: if
you're having to manually over-ride the default, you might as well
explicitly specify it!

Further, so that users will be able to write this common case with a
minimum of boilerplace, the default cabal-version has been bumped to
2.2.

This also brings us into alignment with Cabal 2.2's default
build-type, which is Nothing.