Implement --print-default and generate dhall/defaults/*.dhall  (#105)
`cabal-to-dhall` has recently learned the default values of
fields. Now `dhall-to-cabal` is able to output the default values for
each type.

It's error-prone to keep two sources of defaults; they'll inevitably
drift and disagree. The solution here is to derive one from the other,
and going from a Haskell description of the defaults to Dhall files is
quite possible. `dhall-to-cabal-meta` is a new binary that can
generate the `dhall/defaults/*.dhall` files.

* Use prelude for the default executable scope

This is tidier than building the enumeration each time, and more in
line with how we do other unions.

* defaults: use types.SourceRepo instead of the record directly

* Knock the directory dep down to what shipped with GHC 8.2.