Fully support the mixins field. (#119)
Previously, only the `ModuleRenaming` case was handled. However, there
are two other cases: `DefaultRenaming` and `HidingRenaming`. This
commit adds support for both of those, reducing one more place where
`cabal-to-dhall` can crash.

On the Dhall side, types/ModuleRenaming.dhall is now a union that
closely matches its Cabal-side reflection. The constructors of this
union are exported through the prelude for convenience.

There may be some room for additional convenience functions in Dhall
here, but I'm happy with just having the functionality available.