Add command-line flags to control where the generated file goes. (#113)
Now it's easy to generate a `.cabal` file with the expected name,
instead of having to manually insert the destination name into a
command-line redirection. This is the 99% case, and the remaining 1%
can use `--output-stdout` to pipe the generated file wherever they want.

There are two variants introduced: `--output-dir-cwd DIR` and
`--output-dir-input DIR`. The former interprets its argument relative
to the current working directory, and the latter relative to the input
file, or the current working directory if input is from standard input.

The default is now `--output-dir-input .`.

An example whose brevity is enabled by this:

```
find packages/ -type f -path 'packages/*/package.dhall' -print0 \
  | xargs -0 -p -n1 dhall-to-cabal
```

Without `--output-dir-input`, this requires manual directory movement
and possibly some munging, but now it's easy.

This feature is inspired by hpack, which has similar default behaviour.