Use Dhall.Lint to simplify --print-type and --print-default.
Unconditionally adding a binding for each CSEd subexpression, whether
it was lifted out or not, means less branches and less possibilities
to get things wrong (see #157 for such an example). There's also no
need to keep track of precisely what we've lifted out with CSEState
any more.

In passing, I also spotted a bug in how the CSE code was handling
`let`: each binding applies to subsequent bindings, so we may need to
shift the variable for each one. This doesn't actually happen in
practice, but it's nice to get it right.

None of this ought to be user-visible and, if it is, it's a strict
improvement.