Remove CustomSetup in favour of SetupBuildInfo.
The files are identical, and the decider in favour of doing it this
way around is that SBI was used in slightly more places than CS.

In *theory* it makes sense to separate the buildinfo-ish parts from
the other parts, but in practice the latter set is null, and there's
unlikely to ever be something that'll work similarly enough to
custom-setup that it'll warrant the sharing.

squash! Remove CustomSetup in favour of SetupBuildInfo.